Add --force option to force install package(s)
Sometimes we want to install a package from a new repository and when
doing that, the script will detect that the package is already installed
and won't install the new version of the package.

Adding an option to force the installation of packages (when we know
we've added a new repository and that the package might already have
been installed from elsewhere).